Move java-related code to the azure-intellij-plugin-redis-java library
