+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-redis-java/build.gradle
@@ -1,6 +1,13 @@
 dependencies {
+    compileOnly project(path: ':azure-intellij-plugin-lib')
+    runtimeOnly project(path: ':azure-intellij-plugin-lib', configuration: 'instrumentedJar')
     compileOnly project(path: ':azure-intellij-plugin-redis')
     runtimeOnly project(path: ':azure-intellij-plugin-redis', configuration: 'instrumentedJar')
+    compileOnly project(path: ':azure-intellij-resource-connector-lib')
+    runtimeOnly project(path: ':azure-intellij-resource-connector-lib', configuration: 'instrumentedJar')
     compileOnly project(path: ':azure-intellij-resource-connector-lib-java')
     runtimeOnly project(path: ':azure-intellij-resource-connector-lib-java', configuration: 'instrumentedJar')
+    implementation 'com.microsoft.azure:azure-toolkit-redis-lib'
+    implementation 'com.microsoft.azure:azure-toolkit-ide-common-lib'
+    implementation 'com.microsoft.azure:azure-toolkit-ide-redis-lib'
 }
