+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-redis-java/src/main/java/com/microsoft/azure/toolkit/intellij/redis/connection/RedisResourcePanel.java
